requirements.txt
discord.py==2.4.0
mcstatus==5.1.2
rich==13.5.2